+++ netpyne/batch/batch.py
@@ -69,6 +69,7 @@ def runJob(script, cfgSavePath, netParamsSavePath):
 # -------------------------------------------------------------------------------
 def createFolder(folder):
     import os
+                
     if not os.path.exists(folder):
         try:
             os.mkdir(folder)
@@ -201,8 +202,8 @@ class Batch(object):
 
 
     def openFiles2SaveStats(self):
-        stat_file_name = '%s/%s_stats.cvs' %(self.saveFolder, self.batchLabel)
-        ind_file_name = '%s/%s_stats_indiv.cvs' %(self.saveFolder, self.batchLabel)
+        stat_file_name = '%s/%s_stats.csv' %(self.saveFolder, self.batchLabel)
+        ind_file_name = '%s/%s_stats_indiv.csv' %(self.saveFolder, self.batchLabel)
         individual = open(ind_file_name, 'w')
         stats = open(stat_file_name, 'w')
         stats.write('#gen  pop-size  worst  best  median  average  std-deviation\n')
@@ -315,6 +316,8 @@ class Batch(object):
                     simLabel = self.batchLabel+''.join([''.join('_'+str(i)) for i in iComb])
                     jobName = self.saveFolder+'/'+simLabel  
 
+                    sleepInterval = 1
+
                     # skip if output file already exists
                     if self.runCfg.get('skip', False) and glob.glob(jobName+'.json'):
                         print('Skipping job %s since output file already exists...' % (jobName))
@@ -328,8 +331,6 @@ class Batch(object):
                         self.cfg.saveFolder = self.saveFolder
                         cfgSavePath = self.saveFolder+'/'+simLabel+'_cfg.json'
                         self.cfg.save(cfgSavePath)
-                        
-                        sleepInterval = 1
 
                         # hpc torque job submission
                         if self.runCfg.get('type',None) == 'hpc_torque':
@@ -374,7 +375,7 @@ echo $PBS_O_WORKDIR
                             (output, input) = (proc.stdin, proc.stdout)
 
 
-                        # hpc torque job submission
+                        # hpc slurm job submission
                         elif self.runCfg.get('type',None) == 'hpc_slurm':
 
                             # read params or set defaults
@@ -455,6 +456,7 @@ wait
                             pc.submit(runJob, self.runCfg.get('script', 'init.py'), cfgSavePath, netParamsSavePath)
                             
                         else:
+                            print(self.runCfg)
                             print("Error: invalid runCfg 'type' selected; valid types are 'mpi_bulletin', 'mpi_direct', 'hpc_slurm', 'hpc_torque'")
                             import sys
                             sys.exit(0)
@@ -531,11 +533,16 @@ wait
                     # name and path
                     jobName = "gen_" + str(ngen) + "_cand_" + str(candidate_index)
                     jobPath = genFolderPath + '/' + jobName
+
+                    # set initial cfg initCfg
+                    if len(self.initCfg) > 0:
+                        for paramLabel, paramVal in self.initCfg.items():
+                            self.setCfgNestedParam(paramLabel, paramVal)
                     
                     # modify cfg instance with candidate values
                     for label, value in zip(paramLabels, candidate):
-                        self.setCfgNestedParam(label, value)
                         print('set %s=%s' % (label, value))
+                        self.setCfgNestedParam(label, value)
                     
                     #self.setCfgNestedParam("filename", jobPath)
                     self.cfg.simLabel = jobName
@@ -694,6 +701,8 @@ wait
             def generator(random, args):
                 # generate initial values for candidates
                 return [random.uniform(l, u) for l, u in zip(args.get('lower_bound'), args.get('upper_bound'))]
+
+
             # -------------------------------------------------------------------------------
             # Mutator
             # -------------------------------------------------------------------------------
@@ -769,9 +778,9 @@ wait
                 for iworker in range(int(pc.nhost())):
                     pc.runworker()
 
-            ####################################################################
-            #                       Evolution strategy
-            ####################################################################
+            #------------------------------------------------------------------
+            # Evolutionary algorithm method
+            #-------------------------------------------------------------------
             # Custom algorithm based on Krichmar's params
             if self.evolCfg['evolAlgorithm'] == 'custom':
                 ea = EC.EvolutionaryComputation(rand)
@@ -815,10 +824,12 @@ wait
                 ea.topology = swarm.topologies.ring_topology
             
             else:
-                raise ValueError("%s is not a valid strategy" %(self.evolCfg['evolAlgorithm']))
-            ####################################################################
+                raise ValueError("%s is not a valid strategy" % (self.evolCfg['evolAlgorithm']))
+                
             ea.terminator = EC.terminators.generation_termination
             ea.observer = [EC.observers.stats_observer, EC.observers.file_observer]
+
+
             # -------------------------------------------------------------------------------
             # Run algorithm
             # -------------------------------------------------------------------------------